requirements.txt
impacket
python-dotenv
click
ldap3
dnspython